Add no-op to "flush" plugin -> neovim msg buffer
Sometimes, during a repl-log or dismiss-repl-log call, the plugin
reports that it has written the vim_get_buffers (request) message
to the plugin's output stream. However, neovim (compiled in debug
mode doesn't report receiving it.

This is a workaround, not a true fix. Immediately following the
repl-log requests (neovim -> plugin), neovim also sends a no-op
request. The resulting response (to the no-op), from the plugin
appears to "flush" the plugin -> neovim message "buffer".

Also worth mentioning, we are not able to reproduce this issue
during debug, when using a TCP socket connection. It only occurs
when using STDIO. Also tried using stdbuf to reduce buffering of
the streams, adding a .flush call after every message, to no avail.