Reduce blocking
* Use `rpcnotify()`, rather than `rpcrequest()`, since none of our
  plugin opeerations use the return value.
* Plugin sends channel number to neovim as global variable.

Previously, we noticed that randomly, the plugin would write a request
to its output stream, but the plugin would not notice it. This was due
to the use of `rpcrequest()`. If neovim issued a (1) request to the
plugin, and the plugin tried to make a (2) request of neovim before
responding to request (1), neovim would ignore request (2), because it
was blocked. After receiving the response to request (1), and
unblocking, presumably request (2) was sitting in some IO buffer, and
would not be immediately read by neovim, causing delays.